# jobportal-docs/apis/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/jobs": {
      "post": {
        "description": "Create a new job posting (Employer only)",
        "operationId": "JobsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobDto"
              },
              "examples": {
                "softwareEngineer": {
                  "summary": "Software Engineer Position",
                  "value": {
                    "title": "Senior Software Engineer",
                    "description": "We are seeking an experienced Software Engineer with 5+ years of experience in React and Node.js to join our team.",
                    "salary": 150000,
                    "location": "San Francisco, CA"
                  }
                },
                "productManager": {
                  "summary": "Product Manager Position",
                  "value": {
                    "title": "Product Manager",
                    "description": "Looking for a strategic Product Manager to lead our mobile app initiative and drive product roadmap.",
                    "salary": 140000,
                    "location": "New York, NY"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Job created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Only Employers can create jobs"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create Job Posting",
        "tags": [
          "Jobs"
        ]
      },
      "get": {
        "description": "Get a list of all jobs with optional filtering by title and location",
        "operationId": "JobsController_findAll",
        "parameters": [
          {
            "name": "location",
            "required": false,
            "in": "query",
            "description": "Filter by job location",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "title",
            "required": false,
            "in": "query",
            "description": "Filter by job title",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Jobs retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Job"
                  }
                }
              }
            }
          }
        },
        "summary": "List All Jobs",
        "tags": [
          "Jobs"
        ]
      }
    },
    "/jobs/my": {
      "get": {
        "description": "Get all jobs posted by the authenticated employer",
        "operationId": "JobsController_findMyJobs",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Jobs retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Job"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Only Employers can access"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get My Jobs",
        "tags": [
          "Jobs"
        ]
      }
    },
    "/jobs/{id}": {
      "get": {
        "description": "Retrieve details of a specific job",
        "operationId": "JobsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Job ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job retrieved successfully"
          },
          "404": {
            "description": "Job not found"
          }
        },
        "summary": "Get Job Details",
        "tags": [
          "Jobs"
        ]
      }
    },
    "/jobs/{jobId}": {
      "patch": {
        "description": "Update a job posting (Employer only)",
        "operationId": "JobsController_updateJob",
        "parameters": [
          {
            "name": "jobId",
            "required": true,
            "in": "path",
            "description": "Job ID to update",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateJobDto"
              },
              "examples": {
                "partialUpdate": {
                  "summary": "Update Salary Only",
                  "value": {
                    "salary": 160000
                  }
                },
                "fullUpdate": {
                  "summary": "Update All Fields",
                  "value": {
                    "title": "Lead Software Engineer",
                    "description": "Updated description with new requirements",
                    "salary": 170000,
                    "location": "San Francisco, CA"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Job updated successfully"
          },
          "403": {
            "description": "Forbidden - Only job owner can update"
          },
          "404": {
            "description": "Job not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update Job Posting",
        "tags": [
          "Jobs"
        ]
      },
      "delete": {
        "description": "Delete a job posting (Employer only)",
        "operationId": "JobsController_deleteJob",
        "parameters": [
          {
            "name": "jobId",
            "required": true,
            "in": "path",
            "description": "Job ID to delete",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job deleted successfully"
          },
          "403": {
            "description": "Forbidden - Only job owner can delete"
          },
          "404": {
            "description": "Job not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete Job Posting",
        "tags": [
          "Jobs"
        ]
      }
    },
...
    "/applications/job/{jobId}": {
      "get": {
        "description": "Get all applications for a specific job",
        "operationId": "ApplicationsController_applicationsForJob",
        "parameters": [
          {
            "name": "jobId",
            "required": true,
            "in": "path",
            "description": "Job ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Applications retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Application"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get Job Applications",
        "tags": [
          "Applications"
        ]
      }
    },
...
    "/messages/{receiverId}": {
      "post": {
        "description": "Send a message to another user",
        "operationId": "MessagesController_send",
        "parameters": [
          {
            "name": "receiverId",
            "required": true,
            "in": "path",
            "description": "Recipient User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMessageDto"
              },
              "examples": {
                "shortMessage": {
                  "summary": "Short Message",
                  "value": {
                    "content": "Hi, I am interested in the job posting."
                  }
                },
                "longMessage": {
                  "summary": "Long Message",
                  "value": {
                    "content": "Hello, I am very interested in your job posting for a Senior Software Engineer position. I have 7 years of experience..."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Message sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Recipient not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Send Message",
        "tags": [
          "Messages"
        ]
      }
    },
    "/messages/inbox": {
      "get": {
        "description": "Retrieve all messages received by the authenticated user",
        "operationId": "MessagesController_inbox",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Messages retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Message"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get Inbox",
        "tags": [
          "Messages"
        ]
      }
    },
    "/messages/conversation/{userId}": {
      "get": {
        "description": "Get conversation history with a specific user",
        "operationId": "MessagesController_getConversation",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "Other User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Conversation retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get Conversation",
        "tags": [
          "Messages"
        ]
      }
    },
    "/messages/{messageId}": {
      "delete": {
        "description": "Delete a message from your inbox",
        "operationId": "MessagesController_deleteMessage",
        "parameters": [
          {
            "name": "messageId",
            "required": true,
            "in": "path",
            "description": "Message ID to delete",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Message deleted successfully"
          },
          "404": {
            "description": "Message not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete Message",
        "tags": [
          "Messages"
        ]
      }
    },
...
  },
  "info": {
    "title": "Job Portal API",
    "description": "A comprehensive API for managing jobs, applications, users, and messaging. Supports multiple user roles: JobSeeker, Employer, and Admin.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter JWT token",
        "name": "Authorization",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateJobDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Senior Software Engineer"
          },
          "description": {
            "type": "string",
            "example": "We are looking for a Node.js expert..."
          },
          "salary": {
            "type": "number",
            "example": 120000
          },
          "location": {
            "type": "string",
            "example": "Remote"
          }
        },
        "required": [
          "title",
          "description"
        ]
      },
      "Job": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Software Engineer"
          },
          "description": {
            "type": "string",
            "example": "Job description here..."
          },
          "salary": {
            "type": "number",
            "example": 120000
          },
          "location": {
            "type": "string",
            "example": "Remote / New York"
          },
          "employerId": {
            "type": "string",
            "example": "65c123abc..."
          }
        },
        "required": [
          "title",
          "description",
          "salary",
          "location",
          "employerId"
        ]
      },
      "UpdateJobDto": {
        "type": "object",
        "properties": {}
      },
...
      "Application": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "Applied",
              "Viewed",
              "Rejected"
            ],
            "example": "Applied"
          },
          "resumeUrl": {
            "type": "string",
            "example": "https://cloudinary.com/resume.pdf"
          },
          "coverLetter": {
            "type": "string",
            "example": "I am highly motivated..."
          },
          "jobId": {
            "type": "string",
            "example": "65c123abc..."
          },
          "jobSeekerId": {
            "type": "string",
            "example": "65d456def..."
          }
        },
        "required": [
          "status",
          "resumeUrl",
          "coverLetter",
          "jobId",
          "jobSeekerId"
        ]
      },
      "CreateMessageDto": {
        "type": "object",
        "properties": {}
      },
      "Message": {
        "type": "object",
        "properties": {
          "senderId": {
            "type": "string",
            "example": "65c111aaa..."
          },
          "receiverId": {
            "type": "string",
            "example": "65c222bbb..."
          },
          "content": {
            "type": "string",
            "example": "Hello, how are you?"
          }
        },
        "required": [
          "senderId",
          "receiverId",
          "content"
        ]
      }
    }
  }
}